+++ backend/src/app.module.ts
@@ -9,7 +9,6 @@ import { ReportesModule } from './reportes/reportes.module';
 import { EvidenciaIncidenciaModule } from './evidencia-incidencia/evidencia-incidencia.module';
 import { EmpleadoModule } from './empleado/empleado.module';
 import { AuthModule } from './auth/auth.module';
-import { BitacoraGuardiaModule } from './bitacora-guardia/bitacora-guardia.module';
 
 @Module({
   imports: [
@@ -22,7 +21,6 @@ import { BitacoraGuardiaModule } from './bitacora-guardia/bitacora-guardia.modul
     EvidenciaIncidenciaModule,
     EmpleadoModule,
     AuthModule,
-    BitacoraGuardiaModule,
   ],
   controllers: [AppController],
   providers: [AppService],
