+++ backend/src/app.service.ts
@@ -3,6 +3,6 @@ import { Injectable } from '@nestjs/common';
 @Injectable()
 export class AppService {
   getHello(): string {
-    return 'Hello World!';
+    return 'SERVIDOR BACKEND DE CIVITAS FUNCIONANDO CORRECTAMENTE';
   }
 }
