--- backend/src/bitacora-guardia/bitacora-guardia.module.ts
@@ -1,10 +1,0 @@
-import { Module } from '@nestjs/common';
-import { BitacoraGuardiaService } from './bitacora-guardia.service';
-import { BitacoraGuardiaController } from './bitacora-guardia.controller';
-import { PrismaService } from '../prisma/prisma.service';
-
-@Module({
-  controllers: [BitacoraGuardiaController],
-  providers: [BitacoraGuardiaService, PrismaService],
-})
-export class BitacoraGuardiaModule {}
