+++ backend/src/bitacora/bitacora.service.ts
@@ -3,21 +3,33 @@ import {
   NotFoundException,
   ConflictException,
 } from '@nestjs/common';
-import { Prisma } from '@prisma/client';
 import { PrismaService } from '../prisma/prisma.service';
 
 @Injectable()
 export class BitacoraService {
   constructor(private prisma: PrismaService) {}
 
-  // HU-1.9.1: Ver proveedores dentro del residencial
-  async obtenerProveedoresActivos(filters: {
+  // Ver bitacora con filtros de búsqueda, tipo, residencia, fecha, ordenamiento y paginación
+  async obtenerBitacora(filters: {
     search?: string;
     tipo?: string;
+    residencia?: string;
+    fecha_inicio?: string;
+    fecha_fin?: string;
+    ordenar?: string;
     page: number;
     limit: number;
   }) {
-    const { search, tipo, page, limit } = filters;
+    const {
+      search,
+      tipo,
+      residencia,
+      fecha_inicio,
+      fecha_fin,
+      ordenar,
+      page,
+      limit,
+    } = filters;
 
     const where: any = {
       fecha_hora_salida: null,
@@ -33,19 +45,105 @@ export class BitacoraService {
 
     // SEARCH
     if (search) {
-      where.acceso.visitante.nombre = {
-        contains: search,
-        mode: 'insensitive',
+      where.acceso = {
+        ...where.acceso,
+
+        visitante: {
+          ...where.acceso?.visitante,
+
+          nombre: {
+            contains: search,
+            mode: 'insensitive',
+          },
+        },
       };
     }
 
     // FILTRO TIPO
     if (tipo) {
-      where.acceso.visitante.servicio = {
-        tipo_servicio: {
-          nombre: tipo,
+      where.acceso = {
+        ...where.acceso,
+
+        visitante: {
+          ...where.acceso?.visitante,
+
+          servicio: {
+            tipo_servicio: {
+              categoria: tipo,
+            },
+          },
         },
       };
+    }
+
+    // FILTRO RESIDENCIA
+    if (residencia) {
+      where.acceso = {
+        ...where.acceso,
+
+        visitante: {
+          ...where.acceso?.visitante,
+
+          residente: {
+            vivienda: {
+              numero_vivienda: {
+                contains: residencia,
+                mode: 'insensitive',
+              },
+            },
+          },
+        },
+      };
+    }
+
+    // FILTRO FECHAS
+    if (fecha_inicio || fecha_fin) {
+      where.fecha_hora_entrada = {};
+
+      if (fecha_inicio) {
+        where.fecha_hora_entrada.gte = new Date(fecha_inicio);
+      }
+
+      if (fecha_fin) {
+        where.fecha_hora_entrada.lte = new Date(fecha_fin);
+      }
+    }
+
+    // ORDENAMIENTO
+    let orderBy: any = {
+      fecha_hora_entrada: 'desc',
+    };
+
+    switch (ordenar) {
+      case 'antiguos':
+        orderBy = {
+          fecha_hora_entrada: 'asc',
+        };
+        break;
+
+      case 'nombre':
+        orderBy = {
+          acceso: {
+            visitante: {
+              nombre: 'asc',
+            },
+          },
+        };
+        break;
+
+      case 'tipo':
+        orderBy = {
+          acceso: {
+            visitante: {
+              servicio: {
+                tipo_servicio: {
+                  nombre: 'asc',
+                },
+              },
+            },
+          },
+        };
+        break;
     }
 
     const [data, total] = await Promise.all([
@@ -61,6 +159,18 @@ export class BitacoraService {
                 select: {
                   nombre: true,
                   motivo: true,
+                  es_frecuente: true,
+                  id_servicio: true,
+
+                  residente: {
+                    select: {
+                      vivienda: {
+                        select: {
+                          numero_vivienda: true,
+                        },
+                      },
+                    },
+                  },
 
                   servicio: {
                     select: {
@@ -80,11 +190,15 @@ export class BitacoraService {
               },
             },
           },
+
+          guardia: {
+            select: {
+              nombre: true,
+            },
+          },
         },
 
-        orderBy: {
-          fecha_hora_entrada: 'desc',
-        },
+        orderBy,
 
         skip: (page - 1) * limit,
         take: limit,
@@ -149,81 +263,72 @@ export class BitacoraService {
     return salidaRegistrada;
   }
 
-  async obtenerHistorialBitacora(filters: {
-    fechaInicio?: string;
-    fechaFin?: string;
-    nombre?: string;
-    numeroVivienda?: string;
-    tipoPersona?: 'Residente' | 'Visitante' | 'Servicio';
-    page?: number;
-    limit?: number;
-  }) {
-    const { fechaInicio, fechaFin, nombre, numeroVivienda, page = 1, limit = 10 } = filters;
-    const where: Prisma.BitacoraWhereInput = {};
-
-    // Filtros por fecha
-    if (fechaInicio || fechaFin) {
-      where.fecha_hora_entrada = {};
-      if (fechaInicio) where.fecha_hora_entrada.gte = new Date(fechaInicio);
-      if (fechaFin) where.fecha_hora_entrada.lte = new Date(fechaFin);
-    }
-
-    // Búsqueda por nombre de visitante
-    if (nombre) {
-      where.acceso = {
-        visitante: {
-          nombre: { contains: nombre, mode: 'insensitive' },
-        },
-      };
-    }
-
-    // Filtro por propiedad (Vivienda)
-    if (numeroVivienda) {
-      where.acceso = {
-        ...(where.acceso as any), // Preservar filtros previos si existen
-        visitante: {
-          ...(where.acceso as any)?.visitante,
-          residente: {
-            vivienda: {
-              numero_vivienda: numeroVivienda,
-            },
-          },
-        },
-      };
-    }
-
-    const [data, total] = await Promise.all([
-      this.prisma.bitacora.findMany({
-        where,
-        include: {
-          guardia: true,
-          acceso: {
-            include: {
-              usuario: true, // Quién autorizó
-              visitante: {
-                include: {
-                  residente: { include: { vivienda: true } },
-                  servicio: { include: { tipo_servicio: true } },
+  // Detalle de registro en bitácora a partir de su ID
+  async obtenerDetalleRegistro(id: string) {
+    const registro = await this.prisma.bitacora.findUnique({
+      where: { id_bitacora: id },
+      include: {
+        acceso: {
+          include: {
+            visitante: {
+              include: {
+                servicio: {
+                  include: {
+                    tipo_servicio: true,
+                  },
                 },
+              },
+            },
+            usuario: {
+              include: {
+                persona: true,
               },
             },
           },
         },
-        orderBy: { fecha_hora_entrada: 'desc' },
-        skip: (page - 1) * limit,
-        take: limit,
-      }),
-      this.prisma.bitacora.count({ where }),
-    ]);
+        guardia: true,
+      },
+    });
+
+    if (!registro) {
+      throw new Error('Registro no encontrado');
+    }
+
+    const visitante = registro.acceso.visitante;
+
+    let tipoPersona: string;
+    if (visitante.id_servicio) {
+      tipoPersona = 'proveedor';
+    } else if (visitante.es_frecuente) {
+      tipoPersona = 'empleado_domestico';
+    } else {
+      tipoPersona = 'visitante';
+    }
+
+    let metodoAcceso: string;
+    if (visitante.es_frecuente) metodoAcceso = 'lista';
+    else if (registro.acceso.codigo_qr) metodoAcceso = 'QR';
+    else metodoAcceso = 'manual';
+
+    const estado = registro.fecha_hora_salida ? 'salida' : 'entrada';
 
     return {
-      data,
-      meta: {
-        total,
-        page,
-        limit,
-        totalPages: Math.ceil(total / limit),
+      id: registro.id_bitacora,
+      nombre: visitante.nombre,
+      tipo_persona: tipoPersona,
+      residente_asociado: {
+        nombre: registro.acceso.usuario?.persona?.nombre || '-',
+        avatar_url: registro.acceso.usuario?.persona?.url_imagen || null,
       },
+      fecha_entrada: registro.fecha_hora_entrada,
+      fecha_salida: registro.fecha_hora_salida || '-',
+      metodo_acceso: metodoAcceso,
+      guardia_registro: registro.guardia?.nombre || 'No registrado',
+      estado,
+      avatar_url: visitante.url_imagen || null,
+      qr_utilizado: registro.acceso.codigo_qr || null,
+      notas: registro.comentario_salida || null,
+      hora_validacion: registro.fecha_hora_entrada,
     };
   }
 }
